+++ src/ui/TaskListView.ts
@@ -71,6 +71,9 @@ export class TaskListView extends ItemView {
 	}
 
 
+	/* ItemView declares onOpen as returning Promise<void>. There is nothing to
+	   await, but the signature is not ours to change. */
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async onOpen(): Promise<void> {
 		this.contentEl.addClass("task-base-view");
 		this.render();
@@ -335,6 +338,8 @@ export class TaskListView extends ItemView {
 		new CompleteTaskModal(this.app, task, this.plugin.settings, () => this.render()).open();
 	}
 
+	/* See onOpen. */
+	// eslint-disable-next-line @typescript-eslint/require-await
 	async onClose(): Promise<void> {
 		this.contentEl.empty();
 	}